refactor(net): split virtio net device
